Fixed Issue found in 2nd Round of Testing

+++ voip/src/main/java/com/joshtalks/joshskills/voip/state/LeavingState.kt
@@ -12,8 +12,11 @@ import kotlinx.coroutines.*
 // Fired Leave Channel and waiting for Leave Channel Callback
 class LeavingState(val context: CallContext) : VoipState {
     private val TAG = "LeavingState"
-
-    private val scope = CoroutineScope(Dispatchers.IO)
+    private val scope = CoroutineScope(Dispatchers.IO + CoroutineExceptionHandler {
+            coroutineContext, throwable ->
+        Log.d(TAG, "CoroutineExceptionHandler : $throwable")
+        throwable.printStackTrace()
+    })
 
     init {
         Log.d("Call State", TAG)
@@ -30,22 +33,28 @@ class LeavingState(val context: CallContext) : VoipState {
 
     // Handle Events related to Connected State
     private fun observe() {
+        Log.d(TAG, "observe: ")
         scope.launch {
             try {
                 ensureActive()
                 val event = context.getStreamPipe().receive()
                 ensureActive()
                 if (event.what == CALL_DISCONNECTED) {
-                    Log.d(TAG, "observe: Joined Channel --> ${context.channelData.getChannel()}")
+                    context.closePipe()
                     context.updateUIState(uiState = UIState.empty())
                     ensureActive()
                     PrefManager.setVoipState(State.IDLE)
                     context.destroyContext()
-                } else
-                    throw IllegalEventException("In $TAG but received ${event.what} expected $CALL_INITIATED_EVENT")
-                // TODO: Handle Error Case
+                } else {
+                    ensureActive()
+                    throw IllegalEventException("In $TAG but received ${event.what} expected $CALL_DISCONNECTED")
+                }
                 scope.cancel()
-            } catch (e: Exception) {
+            } catch (e: Throwable) {
+                if(e is CancellationException)
+                    throw e
+                if(e is IllegalEventException)
+                    e.printStackTrace()
                 e.printStackTrace()
                 PrefManager.setVoipState(State.IDLE)
                 context.destroyContext()
